Update README.md
This project allowed me to practice the full end to end process of building a Power BI report the way it would be used in a workplace. From structuring the layout to testing user interactions in the Power BI Service I focused on designing a report that’s intuitive accurate and easy for someone else to interpret.  

It strengthened my understanding of how real users view dashboards and how design decisions directly impact the insights they take away.